requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
peewee==3.17.0
pydantic-settings==2.1.0
python-multipart==0.0.6
openpyxl==3.1.2
python-docx==1.1.0
streamlit
